# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Glory Kamples API",
    "description": "API Kamples — Rust + Axum + OpenAPI",
    "contact": {
      "name": "1ndoryu"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/users/{id}/suspend": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "suspend",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Suspendido"
          },
          "401": {
            "description": "no auth"
          },
          "403": {
            "description": "no admin"
          },
          "404": {
            "description": "no encontrado"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SuspendUserRequest": {
        "type": "object",
        "required": [
          "razon"
        ],
        "properties": {
          "hasta": {
            "type": "string",
            "format": "date-time",
            "description": "Si se omite, suspension indefinida.",
            "nullable": true
          },
          "razon": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}